Stack ADT Implementation Files
Here's all the files. .cpp and .h stackADT files are necessary for it to work, testDriver is for characters but .h and .cpp can work for any other data type if it's properly declared in a test driver file. Also the .txt are both input and output files respectively, with myTest.txt being the input file and out.txt being the output file that's generated once the program is run. The screenshot is the most recent test run that was done with characters.